verificaion de token en BackofficeLayout que hace de filtro de seguridad en todas las rutas /backoffice/*
